refactor: rename tasks-board input-provider to input-context
Replaced `input-provider` with `input-context` in tasks-board components to align with naming conventions. Updated all imports, exports, and references to reflect this change.
